修复Rpc在生成代理时，当成员较多时，会直接出现命名空间的bug

+++ src/TouchSocket.WebApi.Swagger/Common/OpenApiSchema.cs
@@ -33,6 +33,6 @@ namespace TouchSocket.WebApi.Swagger
         public OpenApiSchema Items { get; set; }
 
         [JsonProperty("enum")]
-        public int[] Enum { get; set; }
+        public long[] Enum { get; set; }
     }
 }
